refactor: entity -> domain 수정 반영

+++ src/main/java/com/dbmigrater/DBMigrater/domain/legacy/LegacyUserRepository.java
@@ -1,13 +1,13 @@
 package com.dbmigrater.DBMigrater.domain.legacy;
 
-import com.dbmigrater.DBMigrater.domain.legacy.User;
+import com.dbmigrater.DBMigrater.entity.legacy.User;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface LegacyUserRepository extends MongoRepository<User, Long> {
+public interface LegacyUserRepository extends MongoRepository<User, Integer> {
     User findUserByUserId(Long userId);
 
     List<User> findUserByUserIdBetween(Long startId, Long endId);
