+++ src/main/java/com/dbmigrater/DBMigrater/domain/migration/MigrationUserRepository.java
@@ -1,7 +1,7 @@
 package com.dbmigrater.DBMigrater.domain.migration;
 
-import com.dbmigrater.DBMigrater.domain.migration.User;
+import com.dbmigrater.DBMigrater.entity.migration.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MigrationUserRepository extends JpaRepository<User, Long> {
+public interface MigrationUserRepository extends JpaRepository<User, Integer> {
 }
